refactor(srp): extract self-install logic from cli.py to cli_install.py
cli.py (1400 lines) mixed 4 responsibilities (REPL, oneshot dispatch,
self-install, env recovery). Move the self-install layer (9 helpers,
~400 lines) to cli_install.py:
- _user_scripts_dir, _wrapper_target_dir, _pip_run
- _create_venv_with_deile, _link_global_command
- _ensure_scripts_dir_on_path, _prompt_install_mode
- _run_self_install_async, _run_self_install

cli.py keeps re-exports so deile.cli.<fn> remains importable (the
test suite imports 44 symbols from it). main() now lives next to
the REPL/oneshot logic only.

Tests in test_install_functions.py: monkey-patches via
"deile.cli.<fn>" path adjusted to "deile.cli_install.<fn>" where the
re-export was stale.
